Change value type of ECSInteger from UInt32 to Int64.

+++ CSXTool/ECS/Stuff/ECSObject.cs
@@ -8,9 +8,9 @@ namespace CSXTool.ECS.Stuff
 
     public class ECSInteger : ECSObject
     {
-        public uint Value { get; set; }
+        public long Value { get; set; }
 
-        public ECSInteger(uint value)
+        public ECSInteger(long value)
         {
             Value = value;
         }
